zcbor_common.h: Move tags enum into separate file zcbor_tags.h
